requirements.txt
streamlit==1.31.1
altair==5.2.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.3
seaborn==0.13.2
joblib==1.3.2

# AI / LangChain
langchain>=0.2.14
langchain-community>=0.2.12
langchain-groq>=0.1.5
langgraph>=0.2.20

# Vector DB / Embeddings
chromadb==0.4.24
sentence-transformers==2.7.0

# Utils
python-dotenv==1.0.1
openai==1.30.1